[fix/PYOKEMON-14] EventService 단일 클래스로 수정
