+++ event/src/main/java/com/pyokemon/event/repository/VenueRepository.java
@@ -7,6 +7,5 @@ import com.pyokemon.event.entity.Venue;
 
 @Mapper
 public interface VenueRepository {
-  @Select("SELECT venue_id, venue_name, city, street, zipcode, created_at, updated_at FROM tb_venue WHERE venue_id = #{venueId}")
-  Optional<Venue> findById(Long venueId);
+    Optional<Venue> findById(Long venueId);
 }
