+++ event/src/main/resources/mapper/VenueMapper.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.pyokemon.event.repository.VenueRepository">
+    
+    <select id="findById" parameterType="long" resultType="Venue">
+        SELECT 
+            venue_id,
+            venue_name,
+            city,
+            street,
+            zipcode,
+            created_at,
+            updated_at
+        FROM tb_venue 
+        WHERE venue_id = #{venueId}
+    </select>
+    
+</mapper>
